Init work on Proxy v2
